Create outlierpicker.ipynb
This routine is supposed to pick up outlier globular clusters based on the parameter passed to the routine.
It will return the array of global clusters picked up based on the criteria provided. The parameters to be passed are probability of sample values fall within the range considered as normal. location of the property that is to be used. array of globular clusters.